+++ lib/devm.h
@@ -1,0 +1,10 @@
+#ifndef DEVM_H
+#define DEVM_H
+
+#include "main.h"
+#include "maps.h"
+#include "play.h"
+
+extern void	reveal_map(struct game *cur_game, struct worldmap *map, struct player *cur_player);
+
+#endif