+++ lib/harv.h
@@ -1,0 +1,10 @@
+#ifndef HARV_H
+#define HARV_H
+
+#include "main.h"
+#include "maps.h"
+#include "play.h"
+
+extern void	get_harvest_input(struct game *cur_game, struct worldmap *map, struct player *cur_player);
+
+#endif